Ghost Ticket Bot/requirements.txt
discord.py
Flask
psycopg2
redis
python-dotenv
aiohttp
requests
